+++ crates/model/src/instruments/crypto_perpetual.rs
@@ -270,9 +270,6 @@ impl Hash for CryptoPerpetual {
 }
 
 impl Instrument for CryptoPerpetual {
-    fn tick_scheme(&self) -> Option<Ustr> {
-        self.tick_scheme
-    }
     fn into_any(self) -> InstrumentAny {
         InstrumentAny::CryptoPerpetual(self)
     }
@@ -395,6 +392,14 @@ impl Instrument for CryptoPerpetual {
 
     fn taker_fee(&self) -> Decimal {
         self.taker_fee
+    }
+
+    fn tick_scheme(&self) -> Option<Ustr> {
+        self.tick_scheme
+    }
+
+    fn info(&self) -> Option<&Params> {
+        self.info.as_ref()
     }
 
     fn ts_event(&self) -> UnixNanos {