+++ crates/model/src/instruments/currency_pair.rs
@@ -259,9 +259,6 @@ impl Hash for CurrencyPair {
 }
 
 impl Instrument for CurrencyPair {
-    fn tick_scheme(&self) -> Option<Ustr> {
-        self.tick_scheme
-    }
     fn into_any(self) -> InstrumentAny {
         InstrumentAny::CurrencyPair(self)
     }
@@ -349,6 +346,14 @@ impl Instrument for CurrencyPair {
 
     fn min_price(&self) -> Option<Price> {
         self.min_price
+    }
+
+    fn tick_scheme(&self) -> Option<Ustr> {
+        self.tick_scheme
+    }
+
+    fn info(&self) -> Option<&Params> {
+        self.info.as_ref()
     }
 
     fn ts_event(&self) -> UnixNanos {