+++ crates/model/src/instruments/futures_contract.rs
@@ -255,9 +255,6 @@ impl Hash for FuturesContract {
 }
 
 impl Instrument for FuturesContract {
-    fn tick_scheme(&self) -> Option<Ustr> {
-        self.tick_scheme
-    }
     fn into_any(self) -> InstrumentAny {
         InstrumentAny::FuturesContract(self)
     }
@@ -367,6 +364,14 @@ impl Instrument for FuturesContract {
 
     fn min_price(&self) -> Option<Price> {
         self.min_price
+    }
+
+    fn tick_scheme(&self) -> Option<Ustr> {
+        self.tick_scheme
+    }
+
+    fn info(&self) -> Option<&Params> {
+        self.info.as_ref()
     }
 
     fn ts_event(&self) -> UnixNanos {