+++ crates/model/src/instruments/option_contract.rs
@@ -268,9 +268,6 @@ impl Hash for OptionContract {
 }
 
 impl Instrument for OptionContract {
-    fn tick_scheme(&self) -> Option<Ustr> {
-        self.tick_scheme
-    }
     fn into_any(self) -> InstrumentAny {
         InstrumentAny::OptionContract(self)
     }
@@ -380,6 +377,14 @@ impl Instrument for OptionContract {
 
     fn min_price(&self) -> Option<Price> {
         self.min_price
+    }
+
+    fn tick_scheme(&self) -> Option<Ustr> {
+        self.tick_scheme
+    }
+
+    fn info(&self) -> Option<&Params> {
+        self.info.as_ref()
     }
 
     fn ts_event(&self) -> UnixNanos {