+++ crates/model/src/instruments/perpetual_contract.rs
@@ -290,9 +290,6 @@ impl Hash for PerpetualContract {
 }
 
 impl Instrument for PerpetualContract {
-    fn tick_scheme(&self) -> Option<Ustr> {
-        self.tick_scheme
-    }
     fn into_any(self) -> InstrumentAny {
         InstrumentAny::PerpetualContract(self)
     }
@@ -419,6 +416,14 @@ impl Instrument for PerpetualContract {
 
     fn taker_fee(&self) -> Decimal {
         self.taker_fee
+    }
+
+    fn tick_scheme(&self) -> Option<Ustr> {
+        self.tick_scheme
+    }
+
+    fn info(&self) -> Option<&Params> {
+        self.info.as_ref()
     }
 
     fn ts_event(&self) -> UnixNanos {