+++ crates/model/src/instruments/tokenized_asset.rs
@@ -274,9 +274,6 @@ impl Hash for TokenizedAsset {
 }
 
 impl Instrument for TokenizedAsset {
-    fn tick_scheme(&self) -> Option<Ustr> {
-        self.tick_scheme
-    }
     fn into_any(self) -> InstrumentAny {
         InstrumentAny::TokenizedAsset(self)
     }
@@ -359,6 +356,14 @@ impl Instrument for TokenizedAsset {
 
     fn min_price(&self) -> Option<Price> {
         self.min_price
+    }
+
+    fn tick_scheme(&self) -> Option<Ustr> {
+        self.tick_scheme
+    }
+
+    fn info(&self) -> Option<&Params> {
+        self.info.as_ref()
     }
 
     fn ts_event(&self) -> UnixNanos {